+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add the `alignment` property to `ContextSymbol`, via the `setAlignment` and
+  `getAlignment` methods.
+  - Allows to specify custom alignment to each symbol.
+  - The symbol's address must already be aligned to the given custom alignment,
+    otherwise it will be discarded.
+  - This value is stored in `log2`. For example passing `3` to this value will
+    result on an final alignment of `8` bytes.
+
 ## [1.34.2] - 2025-04-25
 
 ### Fixed